+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Master](https://github.com/appwise-labs/AppwiseCore)
 
-* Script: Renamed the podfile script (`generate_project.rb`) to `cocoapods_rome.rb`.
+* Script: Renamed the podfile script (`generate_project.rb`) to `cocoapods_rome.rb` & add some common steps:
+  * Use the `interface_builder_integration` function in `pre_compile` to fix interface builder integration with pre-compiled frameworks (missing public headers).
+  * Use the `force_bitcode` function in `pre_compile` to ensure frameworks are pre-compiled with Bitcode.
 
 ## [1.2.1](https://github.com/appwise-labs/AppwiseCore/releases/tag/1.2.1)